filled out some of the core class functions

+++ CSCN72030-Anemoi/Sensor.cs
@@ -8,6 +8,6 @@ namespace CSCN72030_Anemoi
 {
     class Sensor
     {
-        public int Data;
+        public float Data;
     }
 }
